The collection tree updating was already happening in its own controller. Renamed it to tree controller for clarity.
